+++ test/all_test.dart
@@ -4,6 +4,7 @@ import 'package:timezone/data/latest.dart';
 
 import 'timeseries_test.dart' as timeseries;
 import 'numeric_timeseries_ext_test.dart' as numeric_ext;
+import 'utils_test.dart' as utils;
 
 void main () {
   initializeTimeZones();
@@ -12,4 +13,5 @@ void main () {
   timeseries.runningGroupTests();
   timeseries.timeseriesTests();
   timeseries.windowTest();
+  utils.tests();
 }